third_party/WebKit/LayoutTests/crypto/random-values-limits-expected.txt

Tests the limits of crypto.randomValues.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'crypto' in self is true
PASS 'getRandomValues' in self.crypto is true
PASS crypto.getRandomValues(almostTooLargeArray) did not throw exception.
PASS crypto.getRandomValues(tooLargeArray) threw exception Error: QuotaExceededError: DOM Exception 22.
PASS successfullyParsed is true

TEST COMPLETE
